+++ apps/prairielearn/src/components/PageLayout.html.ts
@@ -138,12 +138,12 @@ export function PageLayout({
             <div class="${sideNavEnabled ? 'app-main' : ''}">
               <div class="${sideNavEnabled ? 'app-main-container' : ''}">
                 ${resLocals.assessment &&
-                resLocals.assessments &&
+                resLocals.course_instance &&
                 AssessmentNavigation({
+                  courseInstanceId: resLocals.course_instance.id,
                   subPage: navContext.subPage,
-                  courseInstance: resLocals.course_instance,
                   assessment: resLocals.assessment,
-                  assessments: resLocals.assessments,
+                  assessmentSet: resLocals.assessment_set,
                 })}
                 ${showContextNavigation
                   ? ContextNavigation({