+++ apps/prairielearn/src/tests/exam.test.ts
@@ -1,5 +1,4 @@
-import { assert } from 'chai';
-import { afterAll, beforeAll, describe, it } from 'vitest';
+import { afterAll, assert, beforeAll, describe, it } from 'vitest';
 
 import * as sqldb from '@prairielearn/postgres';
 
@@ -1763,11 +1762,9 @@ describe('Exam assessment', { timeout: 60_000 }, function () {
     describe(`partial credit test #${iPartialCreditTest + 1}`, function () {
       describe('server', function () {
         it('should shut down', async function () {
-          // pass "this" explicitly to enable this.timeout() calls
           await helperServer.after();
         });
         it('should start up', async function () {
-          // pass "this" explicitly to enable this.timeout() calls
           await helperServer.before()();
         });
       });