+++ apps/prairielearn/src/tests/showClosedAssessmentScore.test.ts
@@ -1,9 +1,11 @@
-import { assert } from 'chai';
-import { afterAll, beforeAll, describe, test } from 'vitest';
+import { afterAll, assert, beforeAll, describe, test } from 'vitest';
 
 import * as sqldb from '@prairielearn/postgres';
 
 import { config } from '../lib/config.js';
+import { selectAssessmentByTid } from '../models/assessment.js';
+import { ensureEnrollment } from '../models/enrollment.js';
+import { selectUserByUid } from '../models/user.js';
 
 import * as helperClient from './helperClient.js';
 import * as helperServer from './helperServer.js';
@@ -31,8 +33,11 @@ describe(
 
     beforeAll(async function () {
       await helperServer.before()();
-      const results = await sqldb.queryOneRowAsync(sql.select_exam9, []);
-      context.assessmentId = results.rows[0].id;
+      const { id: assessmentId } = await selectAssessmentByTid({
+        course_instance_id: '1',
+        tid: 'exam9-disableRealTimeGradingWithholdGrades',
+      });
+      context.assessmentId = assessmentId;
       context.assessmentUrl = `${context.courseInstanceBaseUrl}/assessment/${context.assessmentId}/`;
     });
 
@@ -47,7 +52,8 @@ describe(
     });
 
     test.sequential('enroll the test student user in the course', async () => {
-      await sqldb.queryAsync(sql.enroll_student_in_course, []);
+      const user = await selectUserByUid('student@example.com');
+      await ensureEnrollment({ user_id: user.user_id, course_instance_id: '1' });
     });
 
     test.sequential('visit start exam page', async () => {