+++ apps/prairielearn/src/tests/vitest.testSetup.ts
@@ -12,12 +12,13 @@ beforeAll(async () => {
 
   // We can't use `import` here because this is a TS file and our tooling
   // isn't yet set up to do dynamic imports of `.ts` files.
-  const { config } = await import('../lib/config.js');
+  const { config, resetConfig } = await import('../lib/config.js');
+
+  // Reset the config to its default state to ensure that state doesn't leak
+  // between tests, even in the case of timeouts that would prevent tests from
+  // cleaning up after themselves.
+  resetConfig();
+
   config.workersCount = 2; // explicitly use 2 workers to test parallelism
   config.fileEditorUseGit = true; // test use of git in file editor
-
-  // Allow using `chai-as-promised` in all tests.
-  const chai = await import('chai');
-  const chaiAsPromised = await import('chai-as-promised');
-  chai.use(chaiAsPromised.default);
 });