+++ apps/prairielearn/vitest.config.ts
@@ -1,13 +1,20 @@
 import { join } from 'path';
 
 import { configDefaults, defineConfig } from 'vitest/config';
-import { BaseSequencer, type TestSpecification } from 'vitest/node';
-
+import { BaseSequencer, type TestSpecification, type Vitest } from 'vitest/node';
+import { GithubActionsReporter } from 'vitest/reporters';
 // Vitest will try to intelligently sequence the test suite based on which ones
 // are slowest. However, this depends on cached data from previous runs, which
 // isn't available in CI. So, we manually specify the slowest tests here and
 // use a custom sequencer to always run them first.
-const SLOW_TESTS = ['src/tests/exampleCourseQuestions.test.ts', 'src/tests/cron.test.ts'];
+const SLOW_TESTS = [
+  'src/tests/exampleCourseQuestions.test.ts',
+  'src/tests/fileEditor.test.ts',
+  'src/tests/homework.test.ts',
+  'src/tests/exam.test.ts',
+  'src/tests/accessibility/index.test.ts',
+  'src/tests/cron.test.ts',
+];
 
 class CustomSequencer extends BaseSequencer {
   async sort(files: TestSpecification[]) {
@@ -29,6 +36,28 @@ class CustomSequencer extends BaseSequencer {
   }
 }
 
+/**
+ * GitHub expects that when formatting an error message, the filename is relative to the project root.
+ * Since we run in a Docker container for CI, we need to ensure that the filename matches what GitHub expects.
+ * https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-error-message
+ *
+ * Vitest may add a first-party API to make this easier:
+ * https://github.com/vitest-dev/vitest/issues/8008
+ */
+export class CustomGithubReporter extends GithubActionsReporter {
+  override onInit(ctx: Vitest) {
+    super.onInit(ctx);
+    const origLog = this.ctx.logger.log;
+    this.ctx.logger.log = (...args: any[]) => {
+      if (args.length === 1) {
+        args[0] = args[0].replaceAll(/(file=)\/PrairieLearn\//g, '$1');
+      }
+
+      origLog.call(this.ctx.logger, ...args);
+    };
+  }
+}
+
 // We support running our tests in two modes:
 //
 // - Directly against the source files in `src/`, in which case we rely on
@@ -46,6 +75,7 @@ const isRunningOnDist = process.argv
 
 export default defineConfig({
   test: {
+    reporters: process.env.GITHUB_ACTIONS ? ['default', new CustomGithubReporter()] : ['default'],
     include: isRunningOnDist
       ? [join(import.meta.dirname, 'dist/**/*.test.js')]
       : [...configDefaults.include],